Refactor: scheduler: add function for checking general node health
In contrast to pe__sum_node_health_scores() which calculates a cumulative score
of all health attributes, the new function pe__node_health() takes the node's
general temperature (red/yellow/green).

It checks whether at least one health attribute is yellow or red, which is a
more intuitive way of saying the node's health is yellow or red (otherwise
multiple yellow health attributes added together could make a node's health
status red).

I went with a comparison-like return value (-1 for red, 0 for yellow, +1 for
green).

I didn't go with pcmk__score_red etc. as the return value because it's possible
that some of those values are equal, and for these purposes we want to be able
to distinguish them. For example, pcmk__score_green and pcmk__score_yellow
might both be 0, but if a health attribute has the textual value "yellow" we
want to return yellow instead of green.

Another possibility would be to add pcmk_rc_red etc. and return those, but that
seemed overkill for this one function.