Fix: tools: crm_resource is-managed check should use meta-attribute table
Previously, check_managed() searched the CIB for the uber-parent's is-managed
attribute. That could give incorrect results if is-managed was set on a
different resource in the ancestry chain (e.g. the resource itself for a group
member, or the group for a cloned group), or if there were multiple is-managed
settings (e.g. using rules).

Now, is-managed is checked in rsc->meta, which should be fully evaluated for
inheritance and rules.