Build: COPR: use custom make target instead of tito
ClusterLabs has a COPR project to build every commit against a few OSes,
to help catch errors early (if anyone looks at it).

Previously, this used tito to create the source RPM. However, it's been broken
since at least the main branch was renamed. Replace it with a .copr/Makefile
which is much simpler anyway.

(This isn't entirely working, but at least it gets further.)