Low: tools: Handle a remote node's name differing from its hostname.
If a remote node is configured so that its cluster name differs from its
hostname, crm_attribute will be confused.  Setting an attribute will
fail with an error message about not being able to map a name to a UUID.

Instead of simply calling get_local_host_name, we can ask the controller
for our name.  This will return a mappable value.

Fixes T34.