API: libpe_status: deprecate node_ping
... along with its associated functionality.

A node in the CIB with a type of "ping" (or in practice, anything but "member"
or "remote") is treated as an offline node. Basically it was a way to do
quorum-only nodes long ago. It has never been documented.

The same thing can be accomplished by putting the node permanently into standby
mode, or using a rule-based location constraint that bans all resources from
the node.