Refactor: libcib: Rearrange the internal find_attr function.
In particular, this removes the value parameter and changes the return
type to be an xmlNodePtr.  These changes, plus the addition of a new
handle_multiples function, allow the callers to decide how to deal with
the possibility of there being multiple attributes.

This can occur if there's more than one attribute with a single name.
The other possibility is if no attribute name is passed at all, so
all attributes for a given node are returned.  We'll be able to handle
that differently in a future commit.