Fix: scheduler: do not enforce resource stop on a rejoined node that was the target of a failed migrate_to
Previously given a scenario:

- rscA failed to migrate from node1 to node2 with a failed migrate_to on
node1
- rscA failed to stop on node2
- node2 got fenced and rscA recovered on node1
- node2 rejoined

, rscA would be considered being "multiple-active" and an unnecessary
full recovery would be issued.

With this commit, when node2 rejoins, rather then enforcing a likely
unnecessary stop of rscA on it, we wait for the probe of it to be
complete, then we will have a better idea about what we should do with it.