Refactor: scheduler: pcmk__node_available() optionally checks score
Some callers of pcmk__node_available() also consider a negative score as
making the node unavailable. Previously, pcmk__node_available() never checked
score, so it was up to the caller to do so.

Now, pcmk__node_available() takes a new argument for whether to check score.
The idea is that when we add new calls to it, it forces us to think about
whether score should be considered. (It also simplifies callers who do need
the check.)