Refactor: scheduler: pcmk__node_available() optionally checks guest
When checking what node a clone instance will be promoted on, we checked
whether the instance is allocated to a guest node whose guest resource won't be
active in addition to calling pcmk__node_available().

This will likely be of use to some other callers, so pcmk__node_available() now
has a new argument for whether to check it. Unfortunately unconditionally
enabling the check causes problems, so we will have to go through each caller
at some point to decide which should use the new argument besides the one
existing location.