Refactor: libcrmcommon: Don't require an IPC connection everywhere.
Any of the new attrd IPC API functions that do not require a response
can just pass NULL for the API and instead we'll use a temporary new
connection.  This connection is synchronous and doesn't have any
callback attached, which is why it can only be used with those functions
that do not send a response.