Fix: scheduler: prevent resources from running on multiple nodes due to partial live migration
Together with the previous commits, this one eventually handles
scenarios like (It's easier to produce with cluster property
batch-limit=1):

- rscA is live-migrating from node1 to node2.

- Before migrate_to on node1 returns, put node2 into standby.

- Transition aborts upon return of successful migrate_to on node1. New
  transition is going to stop the rscA on both nodes and start it on
  node1.

- While it is stopping on node1, run something that is going to make
  the transition abort again like:
  crm_resource  --resource rscA --ban --node node2

- Transition aborts upon return of stop on node1.

Now although there's a stop on node1, it's still a partial migration and
rscA is still potentially active on node2. We need to stop it on node2
to prevent it from running on multiple nodes.