Refactor: tools: Don't create an attrd IPC connection unless needed.
Now that we can just pass NULL for the IPC object most of the time, do
that in the tools.  For the most part, the tools just create a
connection one time, do something, and then tear it down without even
waiting for a response.  For the daemons and query connections that
expect a reply, we still need to create a connection manually.