Feature: controller: record CRM feature set as a transient attribute
Set a transient attribute with CRM feature set for each joined node to
allow to determine if all nodes are updated to the same version. This
information is intended to be queried by tools such as crm_mon or Hawk
which can then warn the user about running a mixed-version cluster.