Refactor: scheduler: merge colocation methods
Merge the rsc_colocation_lh() and rsc_colocation_rh() methods into a single
apply_coloc_score() method that takes an additional boolean argument for
whether it's being called for the dependent or the primary.

This unifies argument validation for the various implementations. Colocations
missing a primary or dependent are ignored when unpacking, so CRM_CHECK() is
sufficient.