Feature: libcrmcommon: Add pcmk__attrd_api_update_list.
This function is like pcmk__attrd_api_update, but it takes a list of
attributes and their values instead.  It constructs a single big IPC
message and sends that to the server.  If the server doesn't support a
single message with multiple attributes, then it falls back to doing
each one individually.

For the moment, the server has not been updated to support this so we
are only doing the fallback path.