Refactor: scheduler: add internal equivalent of pe_graph_flags
enum pe_graph_flags is part of the libpe_status public API, but it is only used
internally. Create an equivalent enum to replace it that is private to
libpacemaker, along with set/clear macros.

It doesn't actually have anything to do with graphs, so the naming is improved
to better reflect the purpose.