Low: CTS: Fix a bug in an awk command used in ctslab.
This command gets the node ID from a cluster node based on the corosync
config file.  4a6d187937 introduced a bug where the command can fail
because there's no longer a space between the command and the file name
awk is operating on.  So, add that space back.

No idea what has been broken as a result of this, if anything.