Refactor: scheduler: functionize cancelling a recurring action for stopped role
RecurringOp_Stopped() previously cancelled the action on the node where the
resource will be active, then looped through all nodes in the cluster, skipping
the node where the resource will be active and scheduling the action on the
rest.

Functionizing makes it convenient to call it in the loop, making it easier to
follow.