Log: scheduler: functionize getting a loggable node name
A lot of log messages include a node name. We generally do a good job of
ensuring that the node structure is not null before using it, but it's still
easy to slip up.

Also, while we currently require all nodes to have a name and ID, it's
conceivable that one might not (with a future supported cluster stack, for
example). So, try name, and fall back to ID.