Low: libstonithd: pull exit reason directly from history replies
1b9e289632 made stonith_api_history() pull the exit reason from the fencer's
reply by pulling the full result with stonith__xe_get_result() and using only
the exit reason.

However stonith__xe_get_result() will log an error if no result is present,
which will be the case if the client is on a newer Pacemaker Remote node
proxied to an older fencer that doesn't set a result.

That's not a big deal, but avoid it anyway by pulling the exit reason alone.