Refactor: scheduler: expose instance number comparator in libpacemaker
Modify sort_rsc_id() to be usable with either clones or bundles, and move it
from pcmk_sched_clone.c to pcmk_sched_resource.c to indicate that.

Also, rename it to pcmk__cmp_instance_number() for clarity, add a doxygen
block, add a sanity check on strrchr() results, and simplify the code a bit.