Low: scheduler: don't prefer greater anti-colocation when ordering instances for assignment
cmp_instance_by_colocation() compares two instances' current nodes by a
cumulative colocation score. Previously, if the first instance's node score was
lower, this would be considered a *greater* preference if the score was negative,
and a lesser preference if the score was non-negative.

This was introduced in a7fb3ca8a2, where it seems unrelated to that commit's
fix, and doesn't makes sense. It especially doesn't make sense that the first
instance was treated differently from the second instance.