Fix: scheduler: don't demote on expected node
demote_resource() loops to schedule a demote on every node where the resource
is active.

Previously, we would skip the loop if the requested node was the expected node
for a multiply active resource with multiple-active=stop_unexpected. This would
work as long as no other nodes were scheduled for demotion; if they were, then
that node's loop will still be done, meaning the demotion would get scheduled
on the expected node anyway.