Fix: Use failed action result in rhcs_validate and _get_metadata
If an action failed but has a non-NULL result, get the rc and other
attributes from that result.

This fixes a regression introduced by b441925, in which failure XML
output now contains a CRM_EX_CONNECTED rc instead of the correct one and
does not contain stdout/stderr. That commit caused
services__execute_file() to return a proper rc instead of TRUE. A
non-pcmk_ok bubbled up the call chain causing
internal_stonith_action_execute() to return -ECONNABORTED. Then
rhcs_validate() and _get_metadata() would use this rc instead of the one
attached to the result.

Resolves: RHBZ#2102292

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>