Log: controller: log an info instead of a warning for a stonith/shutdown that is unknown to the new DC
If it happens to be the old DC that has just shut down, the new DC may
bring up this warning anyway. Although the message is more from the new
DC's point of view purely regarding whether the event matches any
running cluster transition, but it's confusing to end users since the
shutdown was scheduled and should be "expected".

Fixes T454