Refactor: daemons: Add initial new IPC handler to schedulerd.
This only adds the unknown request handler, which is registered but not
yet called from anywhere.  Two things about this patch:

(1) It also ACKs the request, which if it were called would cause a
duplicate ACK to be sent.  The dispatch method currently always ACKs
everything.  I'm going to get rid of this in the future and only have
the message handlers do the ACKing, for consistency with pacemakerd.

(2) Currently, unknown messages are logged and ignored.  The new handler
removes the logging message and sets the result.  This is again for
consistency with how pacemakerd handles unknown requests.