Refactor: daemons: Remove likely unused error handling from schedulerd.
Previously, if pcmk__ipc_send_xml failed, the daemon would write the
transition graph to disk and then cause a core dump.  We don't think
anyone is actually using this behavior, nor do we remember ever needing
this file in order to do any debugging.  For that matter, we can't think
of ever even causing this to happen.

So, remove it to simplify IPC message handling.