Refactor: daemons: Move the schedulerd message handling into dispatch.
This moves the main message handling code block out of
process_pe_message and into its dispatch function.  The couple other
checks on the message can also be moved there.  With this done,
process_pe_message no longer needs to exist.

Note that there is one behavior change here.  Previously, if the op were
empty or missing, the client would still receive an ACK.  However, the
check for a missing op is now handled by pcmk__process_request which
simply returns in that case.  Therefore, the ACK is no longer being sent
to the client.

This is in line with what other daemons that have been converted is
doing, but does still represent a change in behavior.