Test: scheduler: add test for resource-with-container change
The problematic situation is essentially:

- Bundle replica 0 is initially assigned to node1
- Bundle replica 1 is initially assigned to remote-node2,
  and its remote connection to node1
- Bundle replica 0 fails

In the next transition, replica 1 will be assigned first instead of replica 0,
since replica 0 is failed.

Previously, the +10000-score colocation of replica 1's connection with its
container would be considered when assigning replica 1. Replica 1 would
therefore be moved to node1 to be with its connection, and replica 0 would be
recovered on remote-node2. Once the moves were complete, replica 0 would be
clean again, and the replicas might swap places again.

Now, the connection-with-container constraint is ignored when assigning
replica 1, so it stays where it is, and replica 0 is recovered in place.