Low: daemons: Fix a bug processing child XML messages in attrd.
Before, we were just calling attrd_client_update again on the whole
top-level node.  The intention here is that we would pass each child to
attrd_client_update in turn.  This has probably not turned up yet
because it requires a cluster with mixed versions of attrd, in addition
to only happening when messages between daemons are sent.  Normal
command line tool usage will not run into this.

This was introduced by 2e9b5ef562.