Feature: tools: Support regexes with crm_attribute --delete and --update.
According to the documentation, this is supported and it probably does
work when there's an attrd that crm_attribute is talking to.  However,
if operating on a local CIB file, this doesn't work because libxml2
doesn't support regular expressions in xpath.

Support for patterns with --query was added in
ef45001bb5705bfc73b8afe83218c1d04a19c750.  This is a follow up to that.