Test: cts-scheduler: avoid deprecated terminology in regression test inputs
This gets rid of all uses of "master" and "slave" in scheduler regression test
inputs in cases where it doesn't change the outcome of the test. Since no
results change, this commit can be quickly skimmed for review.

Most importantly this changes role references, where the old terms are
deprecated. It also includes non-resource IDs.