Fix: testing: Fix link errors occurring in "make coverage".
Certain test directories cannot compile due to link errors regarding
libgconv and __wrap_getenv.  This only happens if coverage is enabled at
configure time, and then only when "make coverage" is run.

The fix is to remove LDFLAGS_WRAP from Makefile.am.  While only a
handful of test directories are failing at the moment, we are safe if we
remove LDFLAGS_WRAP from all those that do not use a wrapped function.
This is more consistent than just removing it from the ones that fail.