Doc: Clusters from Scratch: Improve /etc/hosts discussion
DNS can be an SPOF if /etc/hosts entries are not added and the addr
option is not specified in the pcs cluster setup command.

Also, the example /etc/hosts entries and ping output used
clusterlabs.org as the domain. The installation example uses localdomain
elsewhere, so we make it consistent here.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>