Low: executor: don't finalize a failed monitor initiation twice
After following the code flow more carefully, I found that if
services_action_async() returns FALSE, it has already called cmd_finalize(), so
we do not need to call it again.

This is not really significant because it would only occur on something like a
pipe() or fork() failure, and the only effect would be logging and notifying
the result twice.

Fixes T68