Test: libcrmcommon: Add a special pcmk__output_formatters function.
This function only gets defined if we are building libcrmcommon_test.a.
The purpose is to expose the internal formatters variable via an
accessor function.  Without that variable, there's not much that can
actually be unit tested in output.c.

Individual unit tests will also have to add this as a CFLAG in order to
have the function defined in the header file.