Test: libcrmcommon: Add unit tests for pcmk__register_format.
However, note that there are no tests for the option adding portion.
GOptionGroup is completely opaque to us, which means the only way we
could test that it actually works is to try to parse a command line.
That seems much more like a unit test of command line handling, whereas
these parameters are kind of incidental to the whole formatted output
framework.