Doc: Pacemaker Remote: Update reason for installing pacemaker on guest
Currently we install pacemaker on the guest node so that we can use
ocf:pacemaker:Dummy. There's no need for this. We can just use
ocf:heartbeat:Dummy (and we'll update that here too).

However, pacemaker does provide crm_attribute, which a lot of resource
agents use.

Also use a loop to create the Dummy resources.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>