Doc: Pacemaker Remote: Remote node pcs explanation
Align "How pcs Configures the Remote" with "How pcs Configures the
Guest." In the near future, we can eliminate the duplication by putting
these in a separate chapter (e.g., an appendix) that the guest and
remote sections both point to.

Also slight verbiage edit in resource section. No real need to point
back to the guest resource section or to add a shared section IMO.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>