Doc: Pacemaker Remote: Use nc in troubleshooting section
nc is somewhat simpler and easier to interpret; it's designed for
exactly this task; and it appears to be safe to run even when the guest
or remote node is already connected to the cluster via
pacemaker-remoted. In my testing there was no disruption to the
remote connection when running the `nc /bin/true` command.

In the near future we should combine these into an appendix, since
they're duplicated info.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>