Build: libcrmcommon: Expose internal library functions for unit testing.
If we are building libcrmcommon_test, undefine G_GNUC_INTERNAL so
functions using this macro are globally visible.  Otherwise, the
functions will be hidden and nothing linking to libcrmcommon_test will
be able to see them.  This makes unit testing them impossible.

This only comes up if libcrmcommon_test is a shared library, which is
about to happen.