Build: libcrmcommon: Build the test library as a shared library.
libtool builds the check_LTLIBRARIES as static by default.  In order to
build the shared version, we need to add -rpath and -version-info
ourselves.

Additionally, the $LDFLAGS_WRAP stuff has to get added.  Otherwise, the
library will contain __real_* symbols that are undefined which causes
link errors building the unit test programs.  We need to tell the linker
how to resolve things with these flags both in the shared library and in
the unit tests.