Test: libcrmcommon: Add unit tests for pcmk__output_and_clear_error.
Again, there really isn't much to test in this function.  We can't test
that it exits early because there's no way to unit test a function that
returns nothing and does nothing.  We can't test that it prints to
fprintf without mocking that function up, which is something I'm not yet
ready to tackle.  So we're really just left with making sure it does
what it's supposed to.