Doc: Pacemaker Remote: Add VirtualDomain resource step
Add step to create VirtualDomain resource. The `pcs cluster node
add-guest` command does **not** do this!

Then add a rule to prevent Pacemaker from starting or probing the
VirtualDomain resource on any node except the node where the VM exists.

Add a slight clarification to start Pacemaker on the host "if it's not
already running." Doesn't fit neatly into another commit and it's next
to this one.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>