Test: cts-exec, cts-fencing: Exit if corosync running; helper function
Exit cts-fencing with an error if corosync is running and --no-cpg is
not set.

Add a helper function to exit with an error message in case any given
cluster process is running, and use that in cts-fencing and cts-exec.

Questionable whether to check for pacemakerd and/or other
other pacemaker daemons but this'll do for now.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>