Doc: libcrmcommon: pcmk__assert_asserts() accepts a statement
Clarify in doc comment that pcmk__assert_asserts() accepts a statement
in general, rather than just an expression. An expression is also valid:
it will be used as an expression statement.

This is important because when cmocka's expect_* and will_return macros
are used along with pcmk__assert_asserts(), they must be called within
the child and not within the parent. Otherwise, the values won't be
removed from the queue in the parent, and the test will fail.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>