Test: libcrmcommon: Mock snprintf()
We'll need this in order to test the snprintf() output error code path
in pcmk__env_option() and pcmk__set_env_option().

This commit also introduces a new pattern for mocking functions that
don't require wrapping because they don't need to call __real_X.
snprintf() meets this criterion: we would never call __real_snprintf(),
because we have to pass the variadic arguments as a va_list to
vsnprintf().

This new pattern is to define a function called pcmk__test_$(fn) in
mock.c, add the function prototype to mock_private.h, and add the name
of the function we're mocking to the MOCKED variable in mk/tap.mk.

The motivation is that some functions like snprintf() are at risk of
being redefined.
  - In Ubuntu, if -O2 is set, _FORTIFY_SOURCE is set to 2 by default.
    This causes snprintf() to be replaced by __builtin___snprintf_chk().
    When this happens, __wrap__snprintf() never gets called.
  - In Fedora 36 ppc64le and above, the addition of support for the IEEE
    128-bit floating point type causes snprintf() to be replaced by
    ___ieee128_snprintf(). As with the Ubuntu case, this causes
    __wrap_snprintf() to never be called.

The Ubuntu case is easy to work around by adding -U_FORTIFY_SOURCE to
CFLAGS. The Fedora case could probably be worked around by wrapping
___ieee128_snprintf() and mapping it to __wrap_snprintf(). More special
cases like this may be required in the future if we handle them
individually. An advantage of this commit's approach is that it doesn't
require addressing special cases.

A disadvantage is that it would not work for mocking v* functions like
vsnprintf(), since we would still need to call the "real" function from
within the wrapper. It also won't work if any other types of functions
(e.g., fopen() or getenv()) ever get redefined in glibc.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>