Low: controld: Check all crm_is_true cases for globally-unique
When validating that globally-unique is not set to true, check all the
cases used by crm_is_true(). Currently this includes (case-insensitive)
"true", "on", "yes", "y", "1".

Also correct the meta attribute name -- should use a hyphen, not an
underscore -- and clarify the error message. The attr doesn't have to be
set explicitly to false, since that's the default.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>