Low: daemons: Don't attempt to update the delay of an unknown attr.
The server will eventually log an error message in this case, but we can
go farther and prevent it from even trying by catching this case in
attrd_client_update.  The client will still not know what's happened
because it isn't waiting for the right response, but that can be handled
later.